+++ packages/cheeket/lib/container.ts
@@ -1,8 +1,7 @@
-import Resolver from './resolver';
 import Register from './register';
 import Token from './token';
 import Middleware, { chain, proxy, route, MiddlewareStorage } from './middleware';
-import ResolveProcessor from './resolve-processor';
+import Resolver, { ResolveProcessor } from './resolver';
 import AsyncEventEmitter from './async-event-emitter';
 
 import InternalTokens from './internal-tokens';
