+++ packages/cheeket/lib/index.ts
@@ -1,4 +1,3 @@
-export { default as Resolver } from './resolver';
 export { default as Context } from './context';
 export { default as Register } from './register';
 export { default as Container } from './container';
@@ -20,3 +19,4 @@ export { default as BindStrategy, asObject, asArray } from './binder';
 export * from './scope';
 
 export { default as Middleware, MiddlewareStorage, proxy, route, compose, chain } from './middleware';
+export { default as Resolver, ResolveError } from './resolver';
