+++ packages/cheeket/lib/resolver/index.ts
@@ -1,0 +1,3 @@
+export { default } from './resolver';
+export { default as ResolveError } from './resolve-error';
+export { default as ResolveProcessor } from './resolve-processor';
