+++ packages/cheeket/lib/resolver/resolve-processor.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-shadow */
 
 import Resolver from './resolver';
-import Token from './token';
-import Context from './context';
+import Token from '../token';
+import Context from '../context';
 import ResolveError from './resolve-error';
-import Middleware from './middleware/middleware';
+import Middleware from '../middleware/middleware';
 
 class ResolveProcessor implements Resolver {
   constructor(private readonly middleware: Middleware<unknown>) {}
