+++ packages/cheeket/lib/resolver/resolver.ts
@@ -1,4 +1,4 @@
-import Token from './token';
+import Token from '../token';
 
 interface Resolver {
   resolveOrDefault<T, D>(token: Token<T>, other: D): Promise<T | D>;
